+++ src/frontend/app/shared-tdoc/components/tdoc-profilechart/tdoc-profilechart.component.ts
@@ -1,11 +1,11 @@
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, Input, Output} from '@angular/core';
-import {TourDocRecord} from '../../../../shared/tdoc-commons/model/records/tdoc-record';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {PlatformService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/platform.service';
-import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
 import {AbstractInlineComponent} from '@dps/mycms-frontend-commons/dist/angular-commons/components/inline.component';
 import {StringUtils} from '@dps/mycms-commons/dist/commons/utils/string.utils';
 import {ChartElement} from '../visjs-profilechart/visjs-profilechart.component';
+import {MapContentUtils} from '../../services/map-contentutils.service';
+import {MapDocRecord} from '../../../../shared/tdoc-commons/services/tdoc-data.utils';
 
 @Component({
     selector: 'app-tdoc-profilechart',
@@ -22,7 +22,7 @@ export class TourDocProfileChartComponent extends AbstractInlineComponent {
     public height: string;
 
     @Input()
-    public tdocs: TourDocRecord[];
+    public docRecords: MapDocRecord[];
 
     @Input()
     public showImageTrackAndGeoPos ? = false;
@@ -30,21 +30,21 @@ export class TourDocProfileChartComponent extends AbstractInlineComponent {
     @Output()
     public chartElementsFound: EventEmitter<ChartElement[]> = new EventEmitter();
 
-    constructor(protected cd: ChangeDetectorRef, private contentUtils: TourDocContentUtils, private appService: GenericAppService,
+    constructor(protected cd: ChangeDetectorRef, private contentUtils: MapContentUtils, private appService: GenericAppService,
                 private platformService: PlatformService) {
         super(cd);
     }
 
     renderChart() {
-        if (!this.tdocs) {
+        if (!this.docRecords) {
             this.chartElements = [];
             return;
         }
 
         const tmpList: ChartElement[] = [];
-        for (let i = 0; i < this.tdocs.length; i++) {
-            const record =  this.tdocs[i];
-            for (const chartElement of this.contentUtils.createChartElementForTourDoc(record, StringUtils.calcCharCodeForListIndex(i + 1),
+        for (let i = 0; i < this.docRecords.length; i++) {
+            const record =  this.docRecords[i];
+            for (const chartElement of this.contentUtils.createChartElementForDocRecord(record, StringUtils.calcCharCodeForListIndex(i + 1),
                 this.showImageTrackAndGeoPos)) {
                 tmpList.push(chartElement);
             }