+++ src/frontend/app/shared-tdoc/components/tdoc-profilemap/tdoc-profilemap.component.ts
@@ -1,11 +1,11 @@
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, Input, Output} from '@angular/core';
-import {TourDocRecord} from '../../../../shared/tdoc-commons/model/records/tdoc-record';
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {PlatformService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/platform.service';
 import {MapElement} from '@dps/mycms-frontend-commons/dist/angular-maps/services/leaflet-geo.plugin';
-import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
 import {AbstractInlineComponent} from '@dps/mycms-frontend-commons/dist/angular-commons/components/inline.component';
 import {StringUtils} from '@dps/mycms-commons/dist/commons/utils/string.utils';
+import {MapContentUtils} from '../../services/map-contentutils.service';
+import {MapDocRecord} from '../../../../shared/tdoc-commons/services/tdoc-data.utils';
 
 @Component({
     selector: 'app-tdoc-profilemap',
@@ -22,29 +22,29 @@ export class TourDocProfileMapComponent extends AbstractInlineComponent {
     public height: string;
 
     @Input()
-    public tdocs: TourDocRecord[];
+    public docRecords: MapDocRecord[];
 
     @Input()
-    public showImageTrackAndGeoPos? = false;
+    public showImageTrackAndGeoPos ? = false;
 
     @Output()
     public mapElementsFound: EventEmitter<MapElement[]> = new EventEmitter();
 
-    constructor(protected cd: ChangeDetectorRef, private contentUtils: TourDocContentUtils, private appService: GenericAppService,
+    constructor(protected cd: ChangeDetectorRef, private contentUtils: MapContentUtils, private appService: GenericAppService,
                 private platformService: PlatformService) {
         super(cd);
     }
 
     renderMap() {
-        if (!this.tdocs) {
+        if (!this.docRecords) {
             this.mapElements = [];
             return;
         }
 
         const tmpList: MapElement[] = [];
-        for (let i = 0; i < this.tdocs.length; i++) {
-            const record =  this.tdocs[i];
-            for (const mapElement of this.contentUtils.createMapElementForTourDoc(record, StringUtils.calcCharCodeForListIndex(i + 1),
+        for (let i = 0; i < this.docRecords.length; i++) {
+            const record =  this.docRecords[i];
+            for (const mapElement of this.contentUtils.createMapElementForDocRecord(record, StringUtils.calcCharCodeForListIndex(i + 1),
                 this.showImageTrackAndGeoPos)) {
                 tmpList.push(mapElement);
             }