+++ src/shared/tdoc-commons/model/records/tdoc-record.ts
@@ -56,8 +56,9 @@ import {
 } from './tdoclinkedplaylist-record';
 import {TourDocLinkedPoiRecord, TourDocLinkedPoiRecordFactory, TourDocLinkedPoiRecordValidator} from './tdoclinkedpoi-record';
 import {TourDocMediaMetaRecord, TourDocMediaMetaRecordFactory, TourDocMediaMetaRecordValidator} from './tdocmediameta-record';
+import {MapDocRecord} from '../../services/tdoc-data.utils';
 
-export interface TourDocRecordType extends CommonDocRecordType {
+export interface TourDocRecordType extends CommonDocRecordType, MapDocRecord {
     locId: number;
     locIdParent: number;
     routeId: number;